+++ pages/admin/forms.py
@@ -3,7 +3,6 @@
 from django import forms
 from django.template.defaultfilters import slugify
 from django.utils.translation import ugettext_lazy as _
-from django.contrib.admin import widgets
 
 from pages import settings
 from pages.models import Page, Content
@@ -12,14 +11,13 @@ from pages.widgets import LanguageChoiceWidget
 
 # error messages
 another_page_error = _('Another page with this slug already exists')
-sibling_position_error = _('A sibiling with this slug already exists at the \
+sibling_position_error = _('A sibling with this slug already exists at the \
 targeted position')
 child_error = _('A child with this slug already exists at the targeted \
 position')
-sibling_error = _('A sibiling with this slug already exists')
-sibling_root_error = _('A sibiling with this slug already exists at the root \
+sibling_error = _('A sibling with this slug already exists')
+sibling_root_error = _('A sibling with this slug already exists at the root \
 level')
-
 
 class PageForm(forms.ModelForm):
     """Form for page creation"""
@@ -84,28 +82,42 @@ it must be unique among the other pages of the same level.')
             elif Content.objects.filter(body=slug, type="slug").count():
                 raise forms.ValidationError(another_page_error)
 
+        if settings.PAGE_USE_SITE_ID:
+            if settings.PAGE_HIDE_SITES:
+                site_ids = [settings.SITE_ID]
+            else:
+                site_ids = [int(x) for x in self.data.getlist('sites')]
+            def intersects_sites(sibling):
+                return sibling.sites.filter(id__in=site_ids).count() > 0
+        else:
+            def intersects_sites(sibling):
+                return True
+
         if not settings.PAGE_UNIQUE_SLUG_REQUIRED:
             if target and position:
                 target = Page.objects.get(pk=target)
                 if position in ['right', 'left']:
                     slugs = [sibling.slug() for sibling in
-                                                target.get_siblings()]
+                             target.get_siblings() 
+                             if intersects_sites(sibling)]
                     slugs.append(target.slug())
                     if slug in slugs:
                         raise forms.ValidationError(sibling_position_error)
                 if position == 'first-child':
                     if slug in [sibling.slug() for sibling in
-                                                target.get_children()]:
+                                target.get_children()
+                                if intersects_sites(sibling)]:
                         raise forms.ValidationError(child_error)
             else:
                 if self.instance.id:
                     if (slug in [sibling.slug() for sibling in
                         self.instance.get_siblings().exclude(
                             id=self.instance.id
-                        )]):
+                        ) if intersects_sites(sibling)]):
                         raise forms.ValidationError(sibling_error)
                 else:
                     if slug in [sibling.slug() for sibling in
-                                                        Page.objects.root()]:
+                                Page.objects.root()
+                                if intersects_sites(sibling)]:
                         raise forms.ValidationError(sibling_root_error)
         return slug