+++ pages/models.py
@@ -80,7 +80,7 @@ class Page(models.Model):
     template = models.CharField(_('template'), max_length=100, null=True,
             blank=True)
 
-    delegate_to = models.CharField(_('template'), max_length=100, null=True,
+    delegate_to = models.CharField(_('delegate to'), max_length=100, null=True,
             blank=True)
 
     freeze_date = models.DateTimeField(_('freeze date'),
@@ -91,16 +91,7 @@ class Page(models.Model):
         sites = models.ManyToManyField(Site, default=[settings.SITE_ID], 
                 help_text=_('The site(s) the page is accessible at.'),
                 verbose_name=_('sites'))
-    else:
-        # Total disabling could make site tests fail
-        _sites = [settings.SITE_ID]
-        def _get_sites(self):
-            return self.sites 
-        def _set_sites(self, new_sites):
-            self.sites = new_sites
-        sites = property(_get_sites, _set_sites)
-        
-            
+
     redirect_to_url = models.CharField(max_length=200, null=True, blank=True)
 
     redirect_to = models.ForeignKey('self', null=True, blank=True,
@@ -116,7 +107,7 @@ class Page(models.Model):
     # per instance cache
     _languages = None
     _complete_slug = None
-    
+    _content_dict = None
 
     class Meta:
         """Make sure the default page ordering is correct."""
@@ -145,6 +136,9 @@ class Page(models.Model):
         # let's assume there is no more broken links after a save
         cache.delete(self.PAGE_BROKEN_LINK_KEY % self.id)
         super(Page, self).save(*args, **kwargs)
+        # fix sites many-to-many link when the're hidden from the form
+        if settings.PAGE_HIDE_SITES and self.sites.count() == 0:
+            self.sites.add(Site.objects.get(pk=settings.SITE_ID))
 
     def _get_calculated_status(self):
         """Get the calculated status of the page based on
@@ -176,6 +170,8 @@ class Page(models.Model):
 
         cache.delete(self.PAGE_LANGUAGES_KEY % (self.id))
         self._languages = None
+        self._complete_slug = None
+        self._content_dict = dict()
 
         p_names = [p.name for p in get_placeholders(self.get_template())]
         if 'slug' not in p_names:
@@ -328,7 +324,7 @@ class Page(models.Model):
         """
         Get the :attr:`template <Page.template>` of this page if
         defined or the closer parent's one if defined
-        or :attr:`pages.settings.DEFAULT_PAGE_TEMPLATE` otherwise.
+        or :attr:`pages.settings.PAGE_DEFAULT_TEMPLATE` otherwise.
         """
         if self.template:
             return self.template
@@ -340,7 +336,7 @@ class Page(models.Model):
                 break
 
         if not template:
-            template = settings.DEFAULT_PAGE_TEMPLATE
+            template = settings.PAGE_DEFAULT_TEMPLATE
 
         return template
 
@@ -348,7 +344,7 @@ class Page(models.Model):
         """
         Get the template name of this page if defined or if a closer
         parent has a defined template or
-        :data:`pages.settings.DEFAULT_PAGE_TEMPLATE` otherwise.
+        :data:`pages.settings.PAGE_DEFAULT_TEMPLATE` otherwise.
         """
         template = self.get_template()
         page_templates = settings.get_page_templates()