+++ pages/settings.py
@@ -2,15 +2,28 @@
 """Convenience module that provides default settings for the ``pages``
 application when the project ``settings`` module does not contain
 the appropriate settings."""
-from os.path import join
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 
+url = 'http://packages.python.org/django-page-cms/settings-list.html#%s'
+
+def get_setting(*args, **kwargs):
+    """Get a setting and raise an appropriate user friendly error if
+    the setting is not found."""
+    for name in args:
+        if hasattr(settings, name):
+            return getattr(settings, name)
+    if kwargs.get('raise_error', False):
+        setting_url = url % args[0].lower().replace('_', '-')
+        raise ImproperlyConfigured('Please make sure you specified at '
+            'least one of these settings: %s \r\nDocumentation: %s'
+            % (args, setting_url))
+    return kwargs.get('default_value', None)
+
+
 # The path to default template 
-DEFAULT_PAGE_TEMPLATE = getattr(settings, 'DEFAULT_PAGE_TEMPLATE', None)
-if DEFAULT_PAGE_TEMPLATE is None:
-    raise ImproperlyConfigured('Please make sure you specified a '
-                               'DEFAULT_PAGE_TEMPLATE setting.')
+PAGE_DEFAULT_TEMPLATE = get_setting('PAGE_DEFAULT_TEMPLATE',
+    'DEFAULT_PAGE_TEMPLATE', raise_error=True)
 
 # PAGE_TEMPLATES is a list of tuples that specifies the which templates 
 # are available in the ``pages`` admin.  Templates should be assigned in 
@@ -29,11 +42,8 @@ if DEFAULT_PAGE_TEMPLATE is None:
 #
 # PAGE_TEMPLATES = _get_templates
 
-PAGE_TEMPLATES = getattr(settings, 'PAGE_TEMPLATES', None)
-if (PAGE_TEMPLATES is None and 
-    not (isinstance(PAGE_TEMPLATES, str) or
-         isinstance(PAGE_TEMPLATES, unicode))):
-    PAGE_TEMPLATES = ()
+PAGE_TEMPLATES = get_setting('PAGE_TEMPLATES',
+    default_value=())
 
 # The callable that is used by the CMS
 def get_page_templates():
@@ -77,15 +87,22 @@ PAGE_CONTENT_REVISION = getattr(settings, 'PAGE_CONTENT_REVISION', True)
 #    ('fr-ch', gettext_noop('Swiss french')),
 #    ('en-us', gettext_noop('US English')),
 #)
-PAGE_LANGUAGES = getattr(settings, 'PAGE_LANGUAGES', settings.LANGUAGES)
+
+PAGE_LANGUAGES = get_setting('PAGE_LANGUAGES', raise_error=True)
 
 # Defines which language should be used by default.  If 
 # ``PAGE_DEFAULT_LANGUAGE`` not specified, then project's
 # ``settings.LANGUAGE_CODE`` is used
-PAGE_DEFAULT_LANGUAGE = getattr(settings, 'PAGE_DEFAULT_LANGUAGE', 
-                                settings.LANGUAGE_CODE)
 
-extra = [('can_freeze', 'Can freeze page',)]
+PAGE_DEFAULT_LANGUAGE = get_setting('PAGE_DEFAULT_LANGUAGE',
+    'LANGUAGE_CODE', raise_error=True)
+
+# Extra Page permission for freezing pages and manage languages
+
+extra = [
+    ('can_freeze', 'Can freeze page',),
+    ('can_publish', 'Can publish page',),
+]
 for lang in PAGE_LANGUAGES:
     extra.append(
         ('can_manage_' + lang[0].replace('-', '_'),
@@ -123,6 +140,10 @@ SITE_ID = getattr(settings, 'SITE_ID', 1)
 # framework
 PAGE_USE_SITE_ID = getattr(settings, 'PAGE_USE_SITE_ID', False)
 
+# Set PAGE_HIDE_SITES to make the sites that appear uneditable and only allow
+# editing and creating of pages on the current site
+PAGE_HIDE_SITES = getattr(settings, 'PAGE_HIDE_SITES', False)
+
 # Set PAGE_USE_LANGUAGE_PREFIX to ``True`` to make the ``get_absolute_url``
 # method to prefix the URLs with the language code
 PAGE_USE_LANGUAGE_PREFIX = getattr(settings, 'PAGE_USE_LANGUAGE_PREFIX',
@@ -139,9 +160,10 @@ PAGE_CONTENT_REVISION_EXCLUDE_LIST = getattr(settings,
 PAGE_SANITIZE_USER_INPUT = getattr(settings, 'PAGE_SANITIZE_USER_INPUT', False)
 
 # URL that handles pages media and uses <MEDIA_ROOT>/pages by default.
-_media_url = getattr(settings, "STATIC_URL", settings.MEDIA_URL)
-PAGES_MEDIA_URL = getattr(settings, 'PAGES_MEDIA_URL',
-    join(_media_url, 'pages/'))
+PAGES_MEDIA_URL = get_setting('PAGES_MEDIA_URL')
+if not PAGES_MEDIA_URL:
+    media_url = get_setting('STATIC_URL', 'MEDIA_URL', raise_error=True)
+    PAGES_MEDIA_URL = media_url + 'pages/'
 
 # Hide the slug's of the first root page ie: ``/home/`` becomes ``/``
 PAGE_HIDE_ROOT_SLUG = getattr(settings, 'PAGE_HIDE_ROOT_SLUG', False)