+++ pages/templatetags/pages_tags.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
 """Page CMS page_tags template tags"""
 from django import template
-from django.utils.safestring import SafeUnicode, mark_safe
-from django.utils.translation import ugettext_lazy as _
-from django.template import Template, TemplateSyntaxError
+from django.utils.safestring import SafeUnicode
+from django.template import TemplateSyntaxError
 #from django.forms import Widget, Textarea, ImageField, CharField
 import urllib
+from django.conf import settings
 
-from pages import settings
+from pages import settings as pages_settings
 from pages.models import Content, Page
 from pages.placeholders import PlaceholderNode, ImagePlaceholderNode
 from pages.placeholders import VideoPlaceholderNode
@@ -15,8 +15,7 @@ from pages.placeholders import parse_placeholder
 
 register = template.Library()
 
-
-def get_page_from_string_or_id(page_string, lang):
+def get_page_from_string_or_id(page_string, lang=None):
     """Return a Page object from a slug or an id."""
     if type(page_string) == int:
         return Page.objects.get(pk=int(page_string))
@@ -34,7 +33,7 @@ def _get_content(context, page, content_type, lang, fallback=True):
         return ''
 
     if not lang and 'lang' in context:
-        lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+        lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
 
     page = get_page_from_string_or_id(page, lang)
 
@@ -74,7 +73,7 @@ def pages_menu(context, page, url='/'):
     :param page: the page where to start the menu from.
     :param url: not used anymore.
     """
-    lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+    lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
     page = get_page_from_string_or_id(page, lang)
     path = context.get('path', None)
     site_id = None
@@ -112,7 +111,7 @@ def pages_sub_menu(context, page, url='/'):
     :param page: the page where to start the menu from.
     :param url: not used anymore.
     """
-    lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+    lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
     page = get_page_from_string_or_id(page, lang)
     path = context.get('path', None)
     if page:
@@ -131,11 +130,14 @@ def pages_siblings_menu(context, page, url='/'):
     :param page: the page where to start the menu from.
     :param url: not used anymore.
     """
-    lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+    lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
     page = get_page_from_string_or_id(page, lang)
     path = context.get('path', None)
     if page:
-        root = page.parent
+        if page.parent:
+            root = page.parent
+        else:
+            root = page
         children = root.get_children_for_frontend()
     if 'current_page' in context:
         current_page = context['current_page']
@@ -146,6 +148,7 @@ pages_siblings_menu = register.inclusion_tag('pages/sub_menu.html',
 def pages_admin_menu(context, page, url='', level=None):
     """Render the admin table of pages."""
     request = context['request']
+    can_publish = context.get('can_publish', False)
     
     if "tree_expanded" in request.COOKIES:
         cookie_string = urllib.unquote(request.COOKIES['tree_expanded'])
@@ -155,10 +158,10 @@ def pages_admin_menu(context, page, url='', level=None):
             if page.id in ids:
                 expanded = True
     
-    page_languages = settings.PAGE_LANGUAGES
+    page_languages = pages_settings.PAGE_LANGUAGES
     #has_permission = page.has_page_permission(request)
-    PAGES_MEDIA_URL = settings.PAGES_MEDIA_URL
-    lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+    PAGES_MEDIA_URL = pages_settings.PAGES_MEDIA_URL
+    lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
     LANGUAGE_BIDI = context.get('LANGUAGE_BIDI', False)
 
     return locals()
@@ -194,7 +197,7 @@ show_content = register.inclusion_tag('pages/content.html',
 def show_slug_with_level(context, page, lang=None, fallback=True):
     """Display slug with level by language."""
     if not lang:
-        lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+        lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
 
     page = get_page_from_string_or_id(page, lang)
     if not page:
@@ -221,7 +224,7 @@ def show_absolute_url(context, page, lang=None):
     :param lang: the wanted language (defaults to `settings.PAGE_DEFAULT_LANGUAGE`)
     """
     if not lang:
-        lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+        lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
     page = get_page_from_string_or_id(page, lang)
     if not page:
         return {'content':''}
@@ -235,7 +238,7 @@ show_absolute_url = register.inclusion_tag('pages/content.html',
 def show_revisions(context, page, content_type, lang=None):
     """Render the last 10 revisions of a page content with a list using
         the ``pages/revisions.html`` template"""
-    if not settings.PAGE_CONTENT_REVISION:
+    if not pages_settings.PAGE_CONTENT_REVISION:
         return {'revisions':None}
     revisions = Content.objects.filter(page=page, language=lang,
                                 type=content_type).order_by('-creation_date')
@@ -256,7 +259,7 @@ def pages_dynamic_tree_menu(context, page, url='/'):
     :param page: the current page
     :param url: not used anymore
     """
-    lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+    lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
     page = get_page_from_string_or_id(page, lang)
     request = context['request']
     site_id = None
@@ -283,7 +286,7 @@ def pages_breadcrumb(context, page, url='/'):
     :param page: the current page
     :param url: not used anymore
     """
-    lang = context.get('lang', settings.PAGE_DEFAULT_LANGUAGE)
+    lang = context.get('lang', pages_settings.PAGE_DEFAULT_LANGUAGE)
     page = get_page_from_string_or_id(page, lang)
     request = context['request']
     site_id = None
@@ -297,6 +300,51 @@ pages_breadcrumb = register.inclusion_tag(
 
 
 """Tags"""
+
+class FakeCSRFNode(template.Node):
+    """Fake CSRF node for django 1.1.1"""
+    def render(self, context):
+        return ''
+def do_csrf_token(parser, token):
+    return FakeCSRFNode()
+try:
+    from django.views.decorators.csrf import csrf_protect
+except ImportError:
+    do_csrf_token = register.tag('csrf_token', do_csrf_token)
+
+class GetPageNode(template.Node):
+    """get_page Node"""
+    def __init__(self, page, varname):
+        self.page = page
+        self.varname = varname
+    
+    def render(self, context):
+        page = get_page_from_string_or_id(self.page)
+        context[self.varname] = page
+        return ''
+
+
+def do_get_page(parser, token):
+    """Retrieve a page and insert into the template's context.
+
+    Example::
+
+        {% get_page "news" as news_page %}
+
+    :param page: the page object, slug or id
+    :param name: name of the context variable to store the page in
+    """
+    bits = token.split_contents()
+    if 4 != len(bits):
+        raise TemplateSyntaxError('%r expects 4 arguments' % bits[0])
+    if bits[-2] != 'as':
+        raise TemplateSyntaxError(
+            '%r expects "as" as the second argument' % bits[0])
+    page = bits[1]
+    varname = bits[-1]
+    return GetPageNode(page, varname)
+do_get_page = register.tag('get_page', do_get_page)
+
 
 class GetContentNode(template.Node):
     """Get content node"""
@@ -314,8 +362,9 @@ class GetContentNode(template.Node):
         )
         return ''
 
+
 def do_get_content(parser, token):
-    """Store a content type from a page into a context variable.
+    """Retrieve a Content object and insert it into the template's context.
 
     Example::
     
@@ -413,4 +462,29 @@ def do_videoplaceholder(parser, token):
     name, params = parse_placeholder(parser, token)
     return VideoPlaceholderNode(name, **params)
 register.tag('videoplaceholder', do_videoplaceholder)
+
+def language_content_up_to_date(page, language):
+    """Tell if all the page content has been updated since the last
+    change of the official version (settings.LANGUAGE_CODE)
+    
+    This is approximated by comparing the last modified date of any
+    content in the page, not comparing each content block to its
+    corresponding official language version.  That allows users to
+    easily make "do nothing" changes to any content block when no 
+    change is required for a language.
+    """
+    lang_code = getattr(settings, 'LANGUAGE_CODE', None)
+    if lang_code == language:
+        # official version is always "up to date"
+        return True
+    # get the last modified date for the official version
+    last_modified = Content.objects.filter(language=lang_code,
+        page=page).order_by('-creation_date')
+    if not last_modified:
+        # no official version
+        return True
+    lang_modified = Content.objects.filter(language=language,
+        page=page).order_by('-creation_date')[0].creation_date
+    return lang_modified > last_modified[0].creation_date
+register.filter(language_content_up_to_date)